Do not rely on systemctl if it is failing
When running in a container or chroot, systemctl is not
working, returning a warning message and stopping.
So all these checks for systemctl return unexpected output,
that is adding a system inconsistency. While the expected
result was the same as intended (always rollback on
containers), the way to get this result was just relying
on invalid assumptions (that systemctl should be running
and reporting the status of the system).
The intention for this patch is to check if systemctl
is running properly and only rely on that output if
that works, adding consistency to the check.
Depending on container runtimes, the behaviour of systemctl
error is different. With podman, it returns offline as
status on stdout, so check for that, and return true for
rollback. With cri-o if throws an error in stderr, and
no output in stdout, so detect that and return true for
rollback.