Unapply bootcmdline settings even when grub.cfg does not exits.
Currently when grub.cfg file is not found, the [bootloader] plugin still
modifies /etc/tuned/bootcmdline file.  However, during rollbacks, the settings
applied to that file are not rolled back.  This becomes a problem in
(containerized) environments when grub.cfg file does not exist.  This patch
will make the behaviour more consistent and unapply the settings in
/etc/tuned/bootcmdline even when no grub.cfg is found.

Signed-off-by: Jiri Mencak <jmencak@users.noreply.github.com>