Make load_all_plugins generic
The load_all_plugins method was written specifically for loading
tuning plugins. However, the PluginLoader class is a loader for all
types of plugins - functions, monitors as well as tuning plugins.

So PluginLoader is not the right place for this kind of method. We
could move the method to tuned.plugins.repository.Repository, however
it's easier to just make it generic for all types of plugins.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>