Add a new file access API
The API is meant to be a replacement for commands.read_file() and
commands.write_to_file().

The new API allows us to log each read and write. At the same time, it
allows the callers to react to different errors that the read/write
operation can result in.

It also allows the callers to log more user-friendly messages - no more
messages like "writing to file foobar failed: Permission denied", that
may not be understandable to a regular user. Instead, we can log more
high-level messages like "Failed to set parameter X to Y: Permission
denied". The actual files that were accessed can still be logged using
the API, however these messages are meant to be debug messages. I.e.,
the 'log_func' argument to FileHandler should be something like
log.debug.

The new API also allows us to test its callers in a better way and
more easily, by replacing the FileOperations class with a mock
implementation that doesn't access real files. Also, the API itself is
very nicely testable, see the included tests.

Another advantage of the new API is that it's separate from the
overcrowded 'commands' class which violates the Single Responsibility
Principle.

The old API provided several features that the new API doesn't provide:
1. The 'makedir' argument to write_to_file() that can be used to create
   directories on the path to the file being written. While having this
   functionality is kind of nice, it's not worth polluting the API with
   an extra parameter that is set to True only twice in the entire
   codebase (one occurence being in tests).
2. The 'no_error' argument, which can be used to suppress logging error
   messages. The same effect ((*) and better) can be achieved by not
   providing 'log_error_func' to the new API.
   (*) The old API is kind of silly in that if no_error == True and an
   error happens, debug log messages about reading/writing to the file
   are still logged, however no messages of any kind regarding the error
   that happened are logged.
3. The 'err_ret' argument to read_file which can be used to obtain a
   default value if reading the file fails. The existence of this
   feature is nothing but laziness to write exception handlers IMHO.
   It's not that hard to write exception handlers, and having them
   results in more readable code, so please, let's just drop this
   feature.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>