Move active profile reading/saving to a separate class
Move active profile reading/saving to a separate class. The 'commands'
class, where the functionality previously resided, is overcrowded and
violates the Single Responsibility Principle.

This is also a first step towards creating better tests for the active
profile manipulation methods.

The methods of the class are not static, because we'll add a
FileHandler attribute to the class later on.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>