Add irqbalance plugin
The irqbalance plugin is a replacement for the bash functions
irqbalance_banned_cpus_setup and irqbalance_banned_cpus_clear. It
configures the banned CPUs in /etc/sysconfig/irqbalance. Then it
restarts irqbalance (if and only if it was previously running).

The banned CPUs can be specified as a CPU list. For example:

[irqbalance]
banned_cpus=2,4,9-13

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>

Resolves: rhbz#1784645