add rtentsk plugin
Add plugin to open a socket with SOF_TIMESTAMPING_OPT_TX_SWHW.

This is for avoiding interruptions due to static key IPIs due
to opening socket with timestamping enabled (by opening a
socket ourselves the static key is kept enabled).

Thanks to Jiri Mencak for porting the C code to Python.

Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>