scheduler: added perf_mmap_pages option to set perf buffer size
For perf events mmapped buffer is used. Under heavy load the buffer may
overflow. In such cases the `scheduler` plugin may start missing events
and not process some newly created processes. Increasing the buffer size
may help.

The buffer size can be set with the `perf_mmap_pages` parameter.
Value of this parameter has to be power of 2. If it is not the power of 2,
the nearest bigger power of 2 value is calculated from it and this
calculated value is used. If the `perf_mmap_pages` parameter is omitted,
the default kernel value is used, which should be 128 for the recent
kernels (tested on kernel-5.9.8).

Example:

[scheduler]
perf_mmap_pages = 256

Resolves: rhbz#1890219

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>