realtime: added support for netdev_queue_count
This allows the user to adjust the queue count for each
network device.

Also add get_queue_count built-in function.

The functions check if the user has specified any
queue count value or not. If no value is passed the function
returns the number of housekeeping CPUs as the queue count.

Resolves: rhbz#1846767

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>