Lower the threading.Event's wait() timeout.
The current 3600s wait() timeout in the controller's loop is too high.
There are known instances of signals not being acted upon.  One known
example is SIGTERM not being acted upon during system shutdown.
Lowering the timeout to 10s allows for this signal to be acted upon right
after the timeout.

Signed-off-by: Jiri Mencak <jmencak@users.noreply.github.com>