tuned-adm: add --verbose option to verify command
'tuned-adm verify' command validates system settings against the
active tuned(8) profiles. But it does not show detailed validation
results to the user. User is left to sift through the tuned(8) log
file to figure it out.

Adding '--verbose' option to the tuned-adm command addresses
this issue. It enables tuned(8) daemon to collect and return
detailed validation results back to the tuned-adm program.
tuned-adm displays these results on the console.

It enables users to confirm that active tuned(8) profile settings
are effective on the system.

Fixes: RHBZ#1956387