configparser: switch to the read_file(), python-3.11 fix
ConfigParser.readfp() is deprecated since python-3.2 and was finally dropped
in the python-3.11 which caused breakage of the TuneD.

This commit switches from the readfp() to the read_file() while still
keeping backward compatibility with the python-3.2 and lower (e.g. python-2.7).

It utilizes wrapper which tries to call the read_file() first and if it
fails it fallbacks to the readfp().

Resolves: rhbz#2038886

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>