sysctl: add protect_sysctl, use it instead of reapply_sysctl
Without TuneD, sysctls from sysctl.conf(5) are set once early on boot,
typically via systemd-sysctl.service(8) or a similar service.
TuneD, however, with reapply_sysctl=1 (which is default), applies these
settings later on during startup and on other occassions as well.

This is rather unfortunate, because it changes the semantics of
sysctl.conf(5). Without TuneD, the configured values serve as mere
defaults while with TuneD their values are maintained as configured.

This has been reported to break things [1] in case of an user who has
configured net.ipv6.conf.*.disable_ipv6=1 in sysctl.conf and the used a
different tool (NetworkManager) to override it later on, essentially
ending up with tools racing for setting the sysctl.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2136749

Let's be a little less aggresive here and just avoid touching
the explicitly configured sysctls. Preserve the old behavior
for compatibility's sake, but turn it off by default.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>