Add an option to skip `write_to_file` if the content would not change
In some situations it may be desirable to check if a file write is
necessary by comparing the new file content with the existing one
(e.g., when the write always causes an inter-processor interrupt).
This commit adds an option to perform such a check in `write_to_file`.
If it succeeds, the write is skipped.