WIP: dump: added new command
The dump command dumps active profile. It's good for debugging, because
the output is flattened, i.e. it shows the resulting profile after
inheritance (profiles merging) is processed.

It keeps naming conventions - for tuned-adm it's named 'dump' action,
for API it's named 'dump_profile' method.

The 'priority' value is not output, but output is ordered according to
it.

The 'cpuinfo_regex' and 'uname_regex' are not output, but output is
rendered according to it.

Variables are not expanded.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>