Refactor per-plugin dynamic tuning configuration
Add a new global option "dynamic_plugins" which can be
used to globally enable dynamic tuning only for specific
plugins.

Also unify the behavior of the "dynamic" profile option
for plugins with dynamic capabilities. All such plugins
now have this option and it is set to True by default.
Dynamic tuning is then enabled for a specific plugin
if it is allowed both globally and locally.