plugin_scheduler: Adjust error logging in _set_affinity
Do not report an error if a process is defunct even if it
is not bound to specific CPUs: failing to change its affinity
does not have any negative effects.

Refactor the helper function _affinity_changeable to
_ignore_set_affinity_error and make it return True/False
depending on whether the failed affinity change can be
ignored (i.e., not reported as an error).

Do not check for vanished processes twice, the later
check is sufficient.

Resolves: RHEL-46560